210411 자유게시판 계층형 댓글 구현중 및 회원가입 구현 준비 (#23)
* Freeboard CRUD

* paste resources

기존에 만들어 두었던 css와 img 파일들을 붙여넣기 하였음.

* 자유게시판 화면처리 진행중

자유게시판 화면처리(목록, 특정 게시물 보기, 수정 등)
수정시 not null이어야 하는 부분을 입력하지 않았더니 에러가 발생하였음.
<form> 안에 <input type="hidden">으로 값을 넣어주었더니 해결되었음.

* 자유게시판 댓글 CRUD 구현중

* 자유게시판 게시물 삭제

DB에서 삭제하는 것이 아닌, 삭제일자를 저장하고, 목록에서 사라지게 하였음.
삭제 후 뒤로가기를 했을 때 게시물을 찾을 수 없다는 창이 나오도록 하는 것이 향후 과제.

* 자유게시판 페이징 처리

* 자유게시판 검색 처리

* For Error Resolving

에러 해결용

* 무시할 커밋(에러 때문)

* 자유게시판 페이징 및 검색 복구

어떤 문제가 있었는지 자세히는 모르겠지만, 페이징과 검색 처리 구현한 것이 사라졌음. 어떻게 해야 하는지 하루종일 고민하여 결국 다시 했음.

* Update pom.xml

댓글 기능 구현에 필요한 JSON 데이터 처리를 위한 라이브러리 등 추가
- jackson-databind
- jackson dataformat xml
- gson

* 210404 Feature/mypage (#10) by kams1011

* "User Crud"

USER CRUD 구현
USER VO 구현
USER SERVICE구현

* regCenter read구현

내가 쓴 글 가져오기

* mypagehome crud

진행중

* thumb read

찜목록 불러오기 완료

* subscribe read

구독목록 불러오기 구현

* regCenter read

개설자센터 데이터 조회 구현 중입니다.

* posted delete

내가 쓴 글 delete 구현했습니다.

* msg delete

메세지 삭제 구현중입니다.

* 210404 meet / paging 기능 구현 reply 기능 구현중  (#11) by DANA-HYEON

* Meet CRUD/Service 구현

모임 VO 구현
모임 CRUD구현
모임 Service 구현

* register ,get기능 구현

/meet/list.jsp html수정
register 기능 구현
get기능 구현

* meet update/delete

meet update/delete 기능 구현

* paging 구현

paging 구현 완료

* Search 구현 중

Search 구현 중

* meet/search기능 구현

meet/ search기능 구현

* meet / reply 기능 구현중

meet / reply 기능 구현중


* 자유게시물 댓글 구현중

- 답글 버튼 클릭 시, 입력창 등장 구현 완료
- 원댓글 입력 구현 완료
- 댓글 좋아요 구현 중

* 210405 Feature/mypage (#13) by kams1011

* "User Crud"

USER CRUD 구현
USER VO 구현
USER SERVICE구현

* regCenter read구현

내가 쓴 글 가져오기

* mypagehome crud

진행중

* thumb read

찜목록 불러오기 완료

* subscribe read

구독목록 불러오기 구현

* regCenter read

개설자센터 데이터 조회 구현 중입니다.

* posted delete

내가 쓴 글 delete 구현했습니다.

* msg delete

메세지 삭제 구현중입니다.

* receive msg delete

수신메세지 삭제 구현했습니다.

* mypage reply read

내가 쓴 댓글 조회처리 완료

* msg send

메시지보내기 구현


* 210405 meet/reply crud 구현 (#12) by DANA-HYEON

* Meet CRUD/Service 구현

모임 VO 구현
모임 CRUD구현
모임 Service 구현

* register ,get기능 구현

/meet/list.jsp html수정
register 기능 구현
get기능 구현

* meet update/delete

meet update/delete 기능 구현

* paging 구현

paging 구현 완료

* Search 구현 중

Search 구현 중

* meet/search기능 구현

meet/ search기능 구현

* meet / reply 기능 구현중

meet / reply 기능 구현중

* meet/reply 댓글 구현

reply 댓글 croud구현
대댓글 기능 구현 중


* 210406 자유게시판 댓글 구현중

- 댓글 답글/수정버튼 클릭 시 입력창 나오게.
- 댓글 수정 구현 완료
- 댓글 작성 에러난 것 해결
- 스프링 시큐리티 공부(별도 프로젝트 폴더 만들어서 이력X)

* 210407 Feature/mypage (#15) by kams1011

* "User Crud"

USER CRUD 구현
USER VO 구현
USER SERVICE구현

* regCenter read구현

내가 쓴 글 가져오기

* mypagehome crud

진행중

* thumb read

찜목록 불러오기 완료

* subscribe read

구독목록 불러오기 구현

* regCenter read

개설자센터 데이터 조회 구현 중입니다.

* posted delete

내가 쓴 글 delete 구현했습니다.

* msg delete

메세지 삭제 구현중입니다.

* receive msg delete

수신메세지 삭제 구현했습니다.

* mypage reply read

내가 쓴 댓글 조회처리 완료

* msg send

메시지보내기 구현

* checkbox checked

체크박스 클릭시 전부 체크되는것 구현

* mypost link

내가 쓴 글 눌렀을 때 이동하기 구현

* category complete

카테고리 셀렉트 구현했습니다.

* message read

메세지 조회 구현


* 210406 meet category list 연동 구현 중 (#16) by DANA-HYEON

* Meet CRUD/Service 구현

모임 VO 구현
모임 CRUD구현
모임 Service 구현

* register ,get기능 구현

/meet/list.jsp html수정
register 기능 구현
get기능 구현

* meet update/delete

meet update/delete 기능 구현

* paging 구현

paging 구현 완료

* Search 구현 중

Search 구현 중

* meet/search기능 구현

meet/ search기능 구현

* meet / reply 기능 구현중

meet / reply 기능 구현중

* meet/reply 댓글 구현

reply 댓글 croud구현
대댓글 기능 구현 중

* meet/reply and category search

meet/reply 구현 중
meet/category search 구현 중

* meet/category 연동 구현 중

meet/category 연동 구현 중


* 210407 자유게시물 댓글/답글 구현중

- 댓글 삭제 구현 완료(DB에 삭제일자 체크 후, 댓글 조회 안되게)
-  답글일 경우, 들여쓰기 및 화살표 표시 구현 (댓글 순서 아직 완전하지 못함)

참고사항 : FREE_REPLY 테이블에서 FRNO2 컬럼이 필요한지 의문. BUNORDER는 필요없을 것 같았지만, 댓글 순서가 뒤죽박죽이 되어 필요할 듯함.

에러 : 새로고침 안하면 답글이나 댓글수정 후 TypeError 발생하여 답글/수정 버튼 작동하지 않음.
TypeError : Cannot set property 'outerHTML' of undefined

* 210407 meet/대댓글, 카테고리 연동, 페이징 구현 완료/ 조회수 찜수 구현 중 (#19) by DANA-HYEON

* Meet CRUD/Service 구현

모임 VO 구현
모임 CRUD구현
모임 Service 구현

* register ,get기능 구현

/meet/list.jsp html수정
register 기능 구현
get기능 구현

* meet update/delete

meet update/delete 기능 구현

* paging 구현

paging 구현 완료

* Search 구현 중

Search 구현 중

* meet/search기능 구현

meet/ search기능 구현

* meet / reply 기능 구현중

meet / reply 기능 구현중

* meet/reply 댓글 구현

reply 댓글 croud구현
대댓글 기능 구현 중

* meet/reply and category search

meet/reply 구현 중
meet/category search 구현 중

* meet/category 연동 구현 중

meet/category 연동 구현 중

* meet/get/category 이름 가져오기

meet/get.jsp에서 category cname 가져오기 구현

* meet/get/내용 그대로 보이기 구현

meet/get/내용 그대로 보이기 구현

* meet/modify/category 이름 가져오기 구현

meet/modify/category 이름 가져오기 구현

* meet/카테고리별 분류

meet/ 모임 수정 시 카테고리 연동 / 카테고리 수정 기능 완료

* meet/category 연동 기능 구현 완료

meet/카테고리 별 검색 / 조회 / 수정 / 삭제 구현완료

* meet/삭제 기능

* meet/reply 대댓글 구현 완료

meet/reply 대댓글 구현 완료

* meet/조회수 찜수 구현중

meet/조회수 찜수 구현중


* 210407 reply로 다시 바꿈 (#21) by kams1011

* "User Crud"

USER CRUD 구현
USER VO 구현
USER SERVICE구현

* regCenter read구현

내가 쓴 글 가져오기

* mypagehome crud

진행중

* thumb read

찜목록 불러오기 완료

* subscribe read

구독목록 불러오기 구현

* regCenter read

개설자센터 데이터 조회 구현 중입니다.

* posted delete

내가 쓴 글 delete 구현했습니다.

* msg delete

메세지 삭제 구현중입니다.

* receive msg delete

수신메세지 삭제 구현했습니다.

* mypage reply read

내가 쓴 댓글 조회처리 완료

* msg send

메시지보내기 구현

* checkbox checked

체크박스 클릭시 전부 체크되는것 구현

* mypost link

내가 쓴 글 눌렀을 때 이동하기 구현

* category complete

카테고리 셀렉트 구현했습니다.

* message read

메세지 조회 구현

* 게시물 클릭시 해당 게시물 주소로 이동

댓글 제외 구현

* merget test

reply 다시 구현


* 210407 Login 기능 준비

* 다현 맞추기용

* 자유게시판 계층형 댓글 진행중

* 자유게시판 조회수 구현 준비

* 회원가입 구현중

* 로그인 구현 준비

+++ src/main/java/com/honeybee/domain/FreeReplyVO.java
@@ -7,9 +7,9 @@ import lombok.Data;
 @Data
 public class FreeReplyVO { // 자유게시물 댓글
 	
-	private Long frno; // 댓글번호
-	private Long frno2; // 답글 대상이 되는 댓글번호
-	private Long fno; // 댓글이 달린 자유게시물 번호
+	private long frno; // 댓글번호
+	private long frno2; // 답글 대상이 되는 댓글번호
+	private long fno; // 댓글이 달린 자유게시물 번호
 	
 	private String id; // 댓글 작성자
 	private String nick; // 댓글 작성자 닉네임
@@ -18,7 +18,7 @@ public class FreeReplyVO { // 자유게시물 댓글
 	private int thumb; // 댓글 좋아요 수
 	private int layer; // 댓글 깊이
 	private int bundle; // 댓글 묶음
-	private int bunorder; // 댓글 묶음 순서
+	private double bunorder; // 댓글 묶음 순서
 	
 	private Date regdt; // 작성일자
 	private Date uptdt; // 수정일자
